dt.brd

Bradford diffusion tube data set, 2020 to 2025 subset of tube measurements collected at sites classified as valid for roadside sampling using the AQSR sampling scheme.

dt.brd

Format

A (11273x11) 'data.frame' object

site

Site identifier code

site_name

Site name/description

year_of_measurement

Year the measurement was made (numeric, YYYY)

month

Month the measurement was made (character, 3-letter abbreviation)

month_numeric

Month the measurement was made (numeric, 1-12)

latitude

Latitude of site

longitude

Longitude of site

local_authority

Local Authority making these measurements

bias_adjusted_measurement

The bias adjusted NO2 measurement

.start_date

Sampling start date (Date, YYYY-MM-DD)

.end_date

Sampling end date (Date, YYYY-MM-DD)

Source

Bradford City Council

Note

Start and end dates of individual diffusion tube sampling periods (.start_date and .end_date) based of Defra Sampling Calendar, see dt.calendar.

Examples

head(dt.brd)
#>     site
#> 1 DT214A
#> 2 DT214B
#> 3 DT214C
#> 4 DT214B
#> 5 DT214C
#> 6 DT214A
#>                                                                            site_name
#> 1 Post outside 221 Bierley Lane near junction with Rockhill Lane opp layby enterance
#> 2 Post outside 221 Bierley Lane near junction with Rockhill Lane opp layby enterance
#> 3 Post outside 221 Bierley Lane near junction with Rockhill Lane opp layby enterance
#> 4 Post outside 221 Bierley Lane near junction with Rockhill Lane opp layby enterance
#> 5 Post outside 221 Bierley Lane near junction with Rockhill Lane opp layby enterance
#> 6 Post outside 221 Bierley Lane near junction with Rockhill Lane opp layby enterance
#>   year_of_measurement month month_numeric latitude longitude local_authority
#> 1                2022   Apr             4 53.75975  -1.73278        bradford
#> 2                2022   Apr             4 53.75975  -1.73278        bradford
#> 3                2022   Apr             4 53.75975  -1.73278        bradford
#> 4                2022   Aug             8 53.75975  -1.73278        bradford
#> 5                2022   Aug             8 53.75975  -1.73278        bradford
#> 6                2022   Dec            12 53.75975  -1.73278        bradford
#>   bias_adjusted_measurement .start_date  .end_date
#> 1                   15.9818  2022-03-30 2022-05-04
#> 2                   17.1790  2022-03-30 2022-05-04
#> 3                   13.4726  2022-03-30 2022-05-04
#> 4                   18.7124  2022-08-03 2022-08-31
#> 5                   15.6702  2022-08-03 2022-08-31
#> 6                   26.4942  2022-11-30 2023-01-04