dt.gla

Nitrogen dioxide (NO2) diffusion tube data for monitoring locations in London, 2021 to 2023.

dt.gla

Format

A (68640x18) 'data.frame' object

X (m)

Site Easting

Y (m)

Site Northing

Site ID

Site Identifier

Site Name

Site Name

Site Type

Site Classification

Valid data capture for monitoring period %

The data capture as a percentage for monitoring period

Valid data capture %

Valid data capture for monitoring period

Annual mean – raw data

Annual mean, as reported

Annual mean – annualised and bias adjusted

Annual mean, after bias correction

local_authority

The local authority

year

Year tube was deployed/measurement taken

month

Month tube was deployed/measurement taken

measurement

The tube measurement reported for that year and month, (in ug/m3; tbc'd)

.longitude

DTEval Tag

.latitude

DTEval Tag

.start_date

DTEval Tag

.end_date

DTEval Tag

.site_id

DTEval Tag

Source

Greater London Authority / London Data Store

https://data.london.gov.uk/publisher/gla-and-tfl-air-quality (downloaded 2024-12-02)

Details

NO2 DT measurement and meta data sets were downloaded from the London Data Store (see Source), and processed as follows:

Monthly measurements were wide-to-long to make month and measurement columns.

WGS84 Latitude and Longitude coordinate tags were calculated for the data using grey.area::geoConvertBNG2LatLon and assuming X (m) and Y (m) are British National Grid (BNG) Easting and Northing coordinates, respectively.

Note

This data set is from a public source but still being QA'ed/cleared. So, contents may change. So, handle with care, and report any concerns.