dt.york

Nitrogen dioxide (NO2) diffusion tube data for monitoring locations about York, 2005 to 2023.

dt.york

Format

A (63408x15) 'data.frame' object

X.OS

Site Easting

Y.OS

Site Northing

SiteId

Site Identifier

SiteName

Site Name

CalendarYear

Year tube was deployed/measurement taken

month

Month tube was deployed/measurement taken

measurement.ppb

Tube measurement (in ppb)

SiteType

Site classification

InAQMA

Site within the Air Quality Management Area (AQMA)?

Distance.to.Kerb.of.Nearest.Road..m.

Distance to kerb of nearest road (m)

measurement

Tube measurement (in ug/m3)

.longitude

DTEval Tag

.latitude

DTEval Tag

.start_date

DTEval Tag

.end_date

DTEval Tag

.site_id

DTEval Tag

Source

City of York Council / York Open Data

https://data.yorkopendata.org/dataset/diffusion-tubes-data (downloaded 2024-12-02)

Details

NO2 DT measurement and meta data sets were downloaded from York Open Data (see Source), and processed as follows:

Measurement data was wide-to-long converted by month and merged with the meta data.

Monthly measurements, described as "raw data (not bias-corrected) and presented in ppb (parts per billion)", were converted to ug/m3 using:

(Defra; EU method; 20oC and 1013mb)

https://uk-air.defra.gov.uk/reports/cat06/0502160851_Conversion_Factors_Between_ppb_and.pdf

measurement.ppb * 1.9125

WGS84 Latitude and Longitude coordinate tags were calculated for the data using grey.area::geoConvertBNG2LatLon and assuming X.OS and Y.OS are British National Grid (BNG) Easting and Northing coordinates, respectively.

Note

This data set is from a public source but still being QA'ed/cleared. So, contents may change. So, handle with care, and report any concerns.