Fit Diffusion Tube Models

Functions for fitting diffusion tube (DT) data calculations using DTEval.

fitTubeModel(
  data,
  tube = ".value",
  inputs = NULL,
  by = NULL,
  model = NULL,
  simplify = FALSE,
  min.count = -1,
  min.prop = -1,
  new.data = NULL,
  ...
)

fitTubeModel_gam(data, tube, inputs, new.data = NULL, ...)

fitTubeModel_loess(data, tube, inputs, new.data = NULL, ...)

Arguments

data

Data source, typically a data.frame or similar, containing data-series of diffusion tube records.

tube

The name of the data-series (in data) to fit a model to (see model), typically the DT NO2 concentrations (in ug/m3).

inputs

The name(s) of the model inputs.

by

The name(s) of hierarchical grouping terms. These are used to sub-sample the data when build multiple models.

model

The model to fit, by default fitTubeModel_loess

simplify

Option to simplify the data before fitting by averaging common inputs, default FALSE.

min.count

The minimum data count required for a model build. The default, -1, disables this option.

min.prop

The minimum proportion of non-NA data required for a model build. The default, -1, disables this option and values higher than 1 exclude all supplied data.

new.data

The type of prediction to generate. The default, NULL, returns the supplied data with an extra column of the requested predictions added as [tube].pred. The alternative option 'input.ranges' generates a data.frame of the model input ranges.

...

additional arguments, currently passed on to model and used if recognised.

Value

All functions return a data.frame of requested data statistics.

Details

fitTubeModel attempts to fit a model to the supplied inputs.

It and related functions assume that data is data DTEval will recognise as Diffusion Tube data, so either previously tagged tube data or data that is tag-able using a default call of tagTube

Examples

if (FALSE) { # \dontrun{
#basic examples
mod <- fitTubeModel(dt.brd, inputs=c(".latitude", ".longitude"),
                    by =".year", simplify = T)
tubePlot(mod, ".value", ".value.pred", facet=".year")
} # }