Miscellaneous Diffusion Tube Data Handling

Miscellaneous code for routine diffusion tube (DT) data handling with DTEval.

getTubeX(data, x = NULL, ..., test.class = NULL, if.err = "return.null")

checkTubeData(
  data,
  x = NULL,
  ...,
  n.x = -1,
  if.err = "return.null",
  output = "data"
)

Arguments

data

Data source, typically a data.frame or similar, containing data-series of diffusion tube records.

x

The name of a data-series in data or expression to be evaluated, supplied as a character string. See Details below.

...

additional arguments, currently ignored.

test.class

If supplied, the required class of x as a character string.

if.err

The action to take if the function fails: By default, 'return.null' to return a NULL silently, but other options include 'stop' to stop as getTubeX, and 'stop<<FUN>>ARG' to stop with function 'FUN' reporting a problem with argument ARG.

n.x

The maximum number of x that to be checked.

output

Where different outputs are an option, the requested output, e.g. data or report.

Value

getTubeX returns x if it can be evaluated with the supplied data.

getTubeData returns data plus any additions columns required to allow all valid x to be used directly as column names, e.g, in form:

data[, x]

Details

getTubeX attempts to extract x from data (or build it from information in data. It uses with(data, x), so can evaluates terms like:

getTubeX(data, 'factor(y)')

(assuming y is in data or visible from it, and x is a character string)

It is intended for use with a single x term and returns it as a vector.

checkTubeData is a general purpose wrapper for getTubeX that handles multiple x terms. It attempts to evaluate each in turn and returns the supplied data plus any additions required.