Miscellaneous Annual Metrics

Miscellaneous functions for the calculation of annual diffusion tube metrics. By default, these are setup for use with DTEval, so expect tagged data (see tagTube), and are intended to be applied by year and location to provide information about local data coverage (or capture efficiency) independent of co-locational sampling or site re-naming.

tubeAnnualCover(data, tube = ".value", by = ".location", ...)

tubeAnnualTest(data, test = NULL, by = ".location", years = NULL, ...)

Arguments

data

Data source, typically a data.frame or similar, containing data-series of diffusion tube records.

tube

The data column containing the diffusion tube measurements, by default .value.

by

The data column containing the site or local codes to be used when grouping tube values for annual statistic calculation, by default .location.

...

additional arguments, e.g. output and rename which allow users to modify function outputs.

test

(tubeAnnualTest only) A test to be evaluated based on data-series in data, supplied as a character string.

years

(tubeAnnualTest only) The years the test should be applied to if not the full range of the data.

Value

By default tubeAnnualCover returns data with two additional columns, .annual.n the numbers of days with at least one measurement reported for that location in that year, and .annual.n the percentage of days with at least one measurement reported for that location in that year.

By default tubeAnnualTest returns data with one additional column, the logical response to the supplied test, for example a call including the argument test='.annual.n>300' will, assuming your data contains a column called .annual.n, generate a column, TRUE for all case where all .annual.n entries for the same year and location were larger than 300 or FALSE if not.

Details

tubeAnnualCover attempts to calculate the annual data coverage, based of the data .start_date and .end_date tags. By default, it is binned and averaged by .location to reduce issues associated with replicate sampling and site renaming, and calculates two metrics: n, the number of days measurements have been reported at that location in the year the sample was collected; and, pc, the percent of the year that accounts for.

tubeAnnualTest can be used to generate logical conditioning based upon tubeAnnualCover outputs. It uses a test argument, supplied as a string, to generate a column of TRUE, FALSE responses, and can be applied to the full data set, or limited to specific years using the extra argument years.

All functions assume data is data DTEval will recognise as diffusion tube data, i.e., previously tagged (see tagTube).