Miscellaneous Latitude and Longitude Handlers

Miscellaneous functions for use with Latitude and Longitude data. By default, these are setup for use with DTEval, so expects tagged data (see tagTube).

tubeInXYPolygon(data, polygon, ...)

Arguments

data

Data source, typically a data.frame or similar, containing data-series of diffusion tube records.

polygon

Data source or polygon shape file.

...

additional arguments, currently ignored.

Value

By default tubeInXYPolygon returns data with an additional column, .in_polygon that identifies tubes as TRUE (inside polygon bounderies) or FALSE (outside polygon bounderies).

Details

tubeinXYPolygon attempts to identify all records in data that are inside the supplied polygon. It is a wrapper for splancs function inout that uses the latitude and longitude tags of the supplied data, and the x and y columns of the supplied polygon to test data rows for inclusion. However, if you want to use different data or polygon columns these can be reset using lat and lon for the data-point locations and x and y for polygon.

All functions assume data and polygon contain point and polygon source information, respectively, that data is data DTEval will recognise as diffusion tube data, i.e., previously tagged (see tagTube), and that these and the x/y values in polygon are WGS84 format map coordinates.

Examples

dt <- tubeInXYPolygon(dt.brd, caz.brd, rename="in.caz")
calcTubeStat(dt, by="in.caz",
      stat=list(mean=mean(.value, na.rm=TRUE),
      count=length(.value[!is.na(.value)])))
#>   in.caz     mean count
#> 1  FALSE 19.90492  3598
#> 2   TRUE 27.21730  7675