Working with Diffusion Tube Meta Data

Miscellaneous code used to work with diffusion tube (DT) meta data.

addTubeMeta(data, ref = NULL, by = NULL, ...)

checkTubeMeta(data, x = NULL, by = NULL, ...)

extractTubeMeta(data, x = NULL, by = NULL, ...)

extractAndAddTubeMeta(data, x = NULL, by = NULL, ref = NULL, ...)

padTubeMeta(data, x = NULL, by = NULL, ...)

repairTubeMeta(data, x = NULL, by = NULL, options = NULL, ...)

Arguments

data

Data source, typically a data.frame or similar, containing data-series of diffusion tube records.

ref

(For addTubeMeta and extractAndAddTubeMeta only) The meta data source, typically a data.frame or similar containing meta-information associated with the diffusion tube records in data. See Details below.

by

The name of a data-series that can be used as a case identifier for meta information. For most diffusion tube meta data, this is often a site identifier, e.g. a site code or name. If by is not supplied, the function will check for known DTEval tags and data set details, but may still require user input. See Details below.

...

additional arguments, currently ignored.

x

(For most ...TubeMeta functions) The name of a data-series in data or expression to be evaluated, supplied as a character string. See Details below.

options

(for repairTubeMeta only) a vector of valid options for x. These are used to a guide when attempting to repair x/by combinations with multiple values associated. See Details below

Value

These functions are generally intended to be used in the form:

updated.data <- addTubeMeta(dt.data, ref, "[by.name]")

requested.data <- checkTubeMeta(dt.data, "[meta.name]", "[by.name]")

... etc

addTubeMeta returns the supplied data with ref merged at the requested level.

checkTubeMeta returns a data.frame of values associated with found when grouping x data at the requested level. More than one value and/or unexpected values indicate that x may not be meta-data or, if it is, that the meta-data has been corrupted.

extractTubeMeta returns a data.frame of data that looks like meta-data when grouped at the requested level. (NB: data-series have to be unique at the grouping level requested to generate an output with extractTubeMeta. So, any x/by combinations that checkTubeMeta reports as having multiple outputs will be ignored by this function.)

padTubeMeta returns data with the requested fix, if it can be applied.

repairTubeMeta like padTubeMeta, returns the fixed data, if the repair can be made.

Details

addTubeMeta attempts to merge data and ref using by as the common merging-term. It is intended for use with reference meta data or data extracted with from a reliable source using extractTubeMeta.

checkTubeMeta attempts to identified all unique values for an x/by combination. Meta-data is expected to unique at a specific level of aggregation, e.g. all latitude records should be identical for a given sample site. So, finding multiple values and/or unexpected values as options for x suggests the data-series is not meta-data at that level or it has been corrupted. (See also repairTubeMeta below.)

extractTubeMeta attempts to extract data that looks like meta information from the supplied data. Although the function default is to look for known sample identifiers, this grouping term can be set using the addition argument by. The function then extracts all data-series with only one unique non-NA value for case of by. By default, the function tests all non-grouping data-series, but similarly testing/extraction can be limited to specific data-series using x. Please remember that this function extracts data-series that looks like meta information at the by grouping-level. There is no unambiguous test to identify a data-series as meta-data. So, this function needs to be handled with care, especially if you have any concerns about the quality of your data sets.

extractAndAddTubeMeta is a wrapper for extractTubeMeta and addTubeMeta to extract valid meta-information from a source, ref, and add that meta-information to data. similarly, care should be taken using this function.

padTubeMeta attempts to pad the x data-series by replacing any NAs with the first non-NA entry from the same data-series. It is intended for use with meta-data data-series, e.g. a latitude (or longitude column) where value was only entered once. Again, care should be taken using this function.

repairTubeMeta attempts to repair what looks like bad meta-information. It uses an extra argument options, a list of valid options for x, as a reference to rationalise multiple value by subsets.