Diffusion Tube Pro-processing

Pre-processing diffusion tube (DT) data for use with DTEval. Coded methods to standardise DT data collected using different DT sampling procedures or different logging conventions. Tagging makes a version of your tube information in a format DTEval can more easily work with, while retaining your original data in the format you are familiar with.

All tag... functions are intended to be run in the form:

tagged.data <- tagTube(data, ...)

(where ... refers to any additional tagging arguments)

tagTube(data, ...)

tagTubeRequired(data, ...)

tagTubeStartEnd(data, method = -1, force = FALSE, ...)

tagTubeLatLon(data, method = -1, force = FALSE, ...)

tagTubeSampleID(data, method = -1, force = FALSE, ...)

tagTubeValue(data, method = -1, force = FALSE, ...)

tagTubeDate(data, method = 2, force = FALSE, ...)

tagTubeMonth(data, method = 1, force = FALSE, ...)

tagTubeYear(data, method = 1, force = FALSE, ...)

tagTubeLocation(data, method = 1, force = FALSE, ...)

Arguments

data

Data source, typically a data.frame or similar, containing data-series of diffusion tube records.

...

Additional arguments can be used to assign data sources if these are not automatically identified by the tag... functions. See below for further details.

method

The method to apply when testing or modifying data. The default (method=-1) means try all methods in order until one works. See below for further details on different function methods.

force

(logical) By default, these functions do not overwrite existing tags if they are already present. The option force=TRUE forces them to rebuild the requested tag using current setting.

Value

These functions typically return the supplied data modified as requested.

Details

All tag... functions allow you to change information sources when adding tags. DTEval has defaults it is expects, but your files might be different. So, you can identify the specific data columns you would like used using additional arguments, e.g.:

year to assign the calendar year

month to assign the calendar month

start and end to assign the individual DT sampling period directly (start deployment and end collection, respectively). start and end are assumed to be provided in conventional 'YYYY-MM-DD' format, but this can also be modified using data.format and strptime formatting instructions.

lat and lon to assign as latitude and longitude, respectively.

value to assign as the tube measurement

So, for example:

tagTubeLatLon(data, lat="lat", lon="long")

tells tagTubeLatLon to use the data columns named 'lat' and 'long' as the sources when tagging latitude and longitude, respectively.

Function Methods:

Methods with -1 defaults try all available methods in order, until they find a method that works or they run out methods. If you want to use a specific method, or method order, just specify them, e.g. to try to tag tube start and end dates first with method 3 then (if that fails) methods 2:

tagTubeStartEnd(data, method=3:2)

tagTubeStartEnd attempts to build Date class time-stamps for sampling dates. It currently has three methods: method=1 use month and year records to build a nominal calendar monthly sampling record, i.e., first to last day of each month; method=2 use month and year records to build a sampling calendar based on the DEFRA diffusion tube sampling calendar (see dt.calendar); and, method=3 to build directly using user-assigned sampling start and end dates. If sampling dates are built, these are added to data as columns .start_date and .end_date.

tagTubeLatLon identifies sample location. It currently uses one method: method=1 assigning all locations based on identified latitude and longitude. These are then tagged as .latitude and .longitude, respectively, and assumed to be WGS84 (EPSG4326) format coordinates.

tagTubeSampleID identifies samples. It currently uses one method: method=1 assigning all same-location-and-time DTs as replicates, and is added to data as .sample_id.

tagTubeValue identifies the diffusion tube measurements, tagging these as .value.

tagTube is a wrapper that runs tagTubeStartEnd, tagTubeLatLon, tagTubeSampleID and tagTubeValue. Other DTEval functions typically pass your data to tagTube to check for tags and add them if missing. If you are using tagTube or one of these on untagged data and need to specific individual methods for each tagTube... function, prefix the method request, e.g. to use method 2 for tagTubeStartEnd but leave the other functions using default setting when applying tagTube, use:

tagTube(data, startend.method=2)

Other tagTube... functions are less commonly used:

tagTubeDate attempts to assign a representative date to each sampling record (data row). By default, it set the date to the mid-point between sample start and end dates, and is, for example, intended for use when plotting time-series: options: 1 (start date), 2 (default; mid-point date); 3 (end date). Because it uses start and end dates, it runs tagTubeStartEnd if these are not already tagged.

tagTubeMonth and tagTubeYear attempt to assign representative months and years to each sampling record (data row). By default, these extract and format values from dates generated using tagTubedate.

tagTubeLocation attempts to assign a representative location to each sampling record (data row). This is typically made by merging the latitude and longitude tags, and uses tagTubeLatLon to get these if not already tagged.

tagTubeRequired is a variation on tagTube that attempts to add any listed in an optional argument, required.

Examples

tagTube(dt.brd)
#>      site
#> 1  DT214A
#> 2  DT214B
#> 3  DT214C
#> 4  DT214B
#> 5  DT214C
#> 6  DT214A
#> 7  DT214B
#> 8  DT214C
#> 9  DT214A
#> 10 DT214B
#> 11 DT214C
#> 12 DT214A
#> 13 DT214B
#> 14 DT214C
#> 15 DT214A
#> 16 DT214B
#> 17 DT214C
#> 18 DT214A
#> 19 DT214B
#> 20 DT214C
#> 21 DT214A
#> 22 DT214B
#> 23 DT214C
#> 24 DT214A
#> 25 DT214B
#> 26 DT214C
#> 27 DT214A
#> 28 DT214B
#> 29 DT214C
#> 30 DT214A
#> 31 DT214B
#> 32 DT214C
#> 33 DT214B
#> 34 DT214C
#> 35 DT260A
#> 36 DT260B
#> 37 DT260A
#> 38 DT260B
#> 39 DT260C
#> 40 DT260A
#> 41 DT260B
#> 42 DT260C
#> 43 DT260A
#> 44 DT260B
#> 45 DT260C
#> 46 DT260A
#> 47 DT260B
#> 48 DT260C
#> 49 DT260A
#> 50 DT260B
#> 51 DT260C
#> 52 DT260A
#> 53 DT260B
#> 54 DT260C
#> 55 DT260A
#> 56 DT260B
#> 57 DT260C
#> 58 DT260A
#> 59 DT260B
#> 60 DT260C
#> 61 DT260A
#> 62 DT260B
#> 63 DT260C
#> 64 DT260A
#> 65 DT260B
#> 66 DT260C
#>                                                                             site_name
#> 1  Post outside 221 Bierley Lane near junction with Rockhill Lane opp layby enterance
#> 2  Post outside 221 Bierley Lane near junction with Rockhill Lane opp layby enterance
#> 3  Post outside 221 Bierley Lane near junction with Rockhill Lane opp layby enterance
#> 4  Post outside 221 Bierley Lane near junction with Rockhill Lane opp layby enterance
#> 5  Post outside 221 Bierley Lane near junction with Rockhill Lane opp layby enterance
#> 6  Post outside 221 Bierley Lane near junction with Rockhill Lane opp layby enterance
#> 7  Post outside 221 Bierley Lane near junction with Rockhill Lane opp layby enterance
#> 8  Post outside 221 Bierley Lane near junction with Rockhill Lane opp layby enterance
#> 9  Post outside 221 Bierley Lane near junction with Rockhill Lane opp layby enterance
#> 10 Post outside 221 Bierley Lane near junction with Rockhill Lane opp layby enterance
#> 11 Post outside 221 Bierley Lane near junction with Rockhill Lane opp layby enterance
#> 12 Post outside 221 Bierley Lane near junction with Rockhill Lane opp layby enterance
#> 13 Post outside 221 Bierley Lane near junction with Rockhill Lane opp layby enterance
#> 14 Post outside 221 Bierley Lane near junction with Rockhill Lane opp layby enterance
#> 15 Post outside 221 Bierley Lane near junction with Rockhill Lane opp layby enterance
#> 16 Post outside 221 Bierley Lane near junction with Rockhill Lane opp layby enterance
#> 17 Post outside 221 Bierley Lane near junction with Rockhill Lane opp layby enterance
#> 18 Post outside 221 Bierley Lane near junction with Rockhill Lane opp layby enterance
#> 19 Post outside 221 Bierley Lane near junction with Rockhill Lane opp layby enterance
#> 20 Post outside 221 Bierley Lane near junction with Rockhill Lane opp layby enterance
#> 21 Post outside 221 Bierley Lane near junction with Rockhill Lane opp layby enterance
#> 22 Post outside 221 Bierley Lane near junction with Rockhill Lane opp layby enterance
#> 23 Post outside 221 Bierley Lane near junction with Rockhill Lane opp layby enterance
#> 24 Post outside 221 Bierley Lane near junction with Rockhill Lane opp layby enterance
#> 25 Post outside 221 Bierley Lane near junction with Rockhill Lane opp layby enterance
#> 26 Post outside 221 Bierley Lane near junction with Rockhill Lane opp layby enterance
#> 27 Post outside 221 Bierley Lane near junction with Rockhill Lane opp layby enterance
#> 28 Post outside 221 Bierley Lane near junction with Rockhill Lane opp layby enterance
#> 29 Post outside 221 Bierley Lane near junction with Rockhill Lane opp layby enterance
#> 30 Post outside 221 Bierley Lane near junction with Rockhill Lane opp layby enterance
#> 31 Post outside 221 Bierley Lane near junction with Rockhill Lane opp layby enterance
#> 32 Post outside 221 Bierley Lane near junction with Rockhill Lane opp layby enterance
#> 33 Post outside 221 Bierley Lane near junction with Rockhill Lane opp layby enterance
#> 34 Post outside 221 Bierley Lane near junction with Rockhill Lane opp layby enterance
#> 35                                                LP8 Netherlands Ave, near scout hut
#> 36                                                LP8 Netherlands Ave, near scout hut
#> 37                                                LP8 Netherlands Ave, near scout hut
#> 38                                                LP8 Netherlands Ave, near scout hut
#> 39                                                LP8 Netherlands Ave, near scout hut
#> 40                                                LP8 Netherlands Ave, near scout hut
#> 41                                                LP8 Netherlands Ave, near scout hut
#> 42                                                LP8 Netherlands Ave, near scout hut
#> 43                                                LP8 Netherlands Ave, near scout hut
#> 44                                                LP8 Netherlands Ave, near scout hut
#> 45                                                LP8 Netherlands Ave, near scout hut
#> 46                                                LP8 Netherlands Ave, near scout hut
#> 47                                                LP8 Netherlands Ave, near scout hut
#> 48                                                LP8 Netherlands Ave, near scout hut
#> 49                                                LP8 Netherlands Ave, near scout hut
#> 50                                                LP8 Netherlands Ave, near scout hut
#> 51                                                LP8 Netherlands Ave, near scout hut
#> 52                                                LP8 Netherlands Ave, near scout hut
#> 53                                                LP8 Netherlands Ave, near scout hut
#> 54                                                LP8 Netherlands Ave, near scout hut
#> 55                                                LP8 Netherlands Ave, near scout hut
#> 56                                                LP8 Netherlands Ave, near scout hut
#> 57                                                LP8 Netherlands Ave, near scout hut
#> 58                                                LP8 Netherlands Ave, near scout hut
#> 59                                                LP8 Netherlands Ave, near scout hut
#> 60                                                LP8 Netherlands Ave, near scout hut
#> 61                                                LP8 Netherlands Ave, near scout hut
#> 62                                                LP8 Netherlands Ave, near scout hut
#> 63                                                LP8 Netherlands Ave, near scout hut
#> 64                                                LP8 Netherlands Ave, near scout hut
#> 65                                                LP8 Netherlands Ave, near scout hut
#> 66                                                LP8 Netherlands Ave, near scout hut
#>    year_of_measurement month month_numeric latitude longitude local_authority
#> 1                 2022   Apr             4 53.75975  -1.73278        bradford
#> 2                 2022   Apr             4 53.75975  -1.73278        bradford
#> 3                 2022   Apr             4 53.75975  -1.73278        bradford
#> 4                 2022   Aug             8 53.75975  -1.73278        bradford
#> 5                 2022   Aug             8 53.75975  -1.73278        bradford
#> 6                 2022   Dec            12 53.75975  -1.73278        bradford
#> 7                 2022   Dec            12 53.75975  -1.73278        bradford
#> 8                 2022   Dec            12 53.75975  -1.73278        bradford
#> 9                 2022   Feb             2 53.75975  -1.73278        bradford
#> 10                2022   Feb             2 53.75975  -1.73278        bradford
#> 11                2022   Feb             2 53.75975  -1.73278        bradford
#> 12                2022   Jan             1 53.75975  -1.73278        bradford
#> 13                2022   Jan             1 53.75975  -1.73278        bradford
#> 14                2022   Jan             1 53.75975  -1.73278        bradford
#> 15                2022   Jul             7 53.75975  -1.73278        bradford
#> 16                2022   Jul             7 53.75975  -1.73278        bradford
#> 17                2022   Jul             7 53.75975  -1.73278        bradford
#> 18                2022   Jun             6 53.75975  -1.73278        bradford
#> 19                2022   Jun             6 53.75975  -1.73278        bradford
#> 20                2022   Jun             6 53.75975  -1.73278        bradford
#> 21                2022   Mar             3 53.75975  -1.73278        bradford
#> 22                2022   Mar             3 53.75975  -1.73278        bradford
#> 23                2022   Mar             3 53.75975  -1.73278        bradford
#> 24                2022   May             5 53.75975  -1.73278        bradford
#> 25                2022   May             5 53.75975  -1.73278        bradford
#> 26                2022   May             5 53.75975  -1.73278        bradford
#> 27                2022   Nov            11 53.75975  -1.73278        bradford
#> 28                2022   Nov            11 53.75975  -1.73278        bradford
#> 29                2022   Nov            11 53.75975  -1.73278        bradford
#> 30                2022   Oct            10 53.75975  -1.73278        bradford
#> 31                2022   Oct            10 53.75975  -1.73278        bradford
#> 32                2022   Oct            10 53.75975  -1.73278        bradford
#> 33                2022   Sep             9 53.75975  -1.73278        bradford
#> 34                2022   Sep             9 53.75975  -1.73278        bradford
#> 35                2022   Apr             4 53.75981  -1.76838        bradford
#> 36                2022   Apr             4 53.75981  -1.76838        bradford
#> 37                2022   Aug             8 53.75981  -1.76838        bradford
#> 38                2022   Aug             8 53.75981  -1.76838        bradford
#> 39                2022   Aug             8 53.75981  -1.76838        bradford
#> 40                2022   Dec            12 53.75981  -1.76838        bradford
#> 41                2022   Dec            12 53.75981  -1.76838        bradford
#> 42                2022   Dec            12 53.75981  -1.76838        bradford
#> 43                2022   Feb             2 53.75981  -1.76838        bradford
#> 44                2022   Feb             2 53.75981  -1.76838        bradford
#> 45                2022   Feb             2 53.75981  -1.76838        bradford
#> 46                2022   Jan             1 53.75981  -1.76838        bradford
#> 47                2022   Jan             1 53.75981  -1.76838        bradford
#> 48                2022   Jan             1 53.75981  -1.76838        bradford
#> 49                2022   Jul             7 53.75981  -1.76838        bradford
#> 50                2022   Jul             7 53.75981  -1.76838        bradford
#> 51                2022   Jul             7 53.75981  -1.76838        bradford
#> 52                2022   Jun             6 53.75981  -1.76838        bradford
#> 53                2022   Jun             6 53.75981  -1.76838        bradford
#> 54                2022   Jun             6 53.75981  -1.76838        bradford
#> 55                2022   Mar             3 53.75981  -1.76838        bradford
#> 56                2022   Mar             3 53.75981  -1.76838        bradford
#> 57                2022   Mar             3 53.75981  -1.76838        bradford
#> 58                2022   May             5 53.75981  -1.76838        bradford
#> 59                2022   May             5 53.75981  -1.76838        bradford
#> 60                2022   May             5 53.75981  -1.76838        bradford
#> 61                2022   Nov            11 53.75981  -1.76838        bradford
#> 62                2022   Nov            11 53.75981  -1.76838        bradford
#> 63                2022   Nov            11 53.75981  -1.76838        bradford
#> 64                2022   Oct            10 53.75981  -1.76838        bradford
#> 65                2022   Oct            10 53.75981  -1.76838        bradford
#> 66                2022   Oct            10 53.75981  -1.76838        bradford
#>    bias_adjusted_measurement .start_date  .end_date .latitude .longitude
#> 1                    15.9818  2022-03-30 2022-05-04  53.75975   -1.73278
#> 2                    17.1790  2022-03-30 2022-05-04  53.75975   -1.73278
#> 3                    13.4726  2022-03-30 2022-05-04  53.75975   -1.73278
#> 4                    18.7124  2022-08-03 2022-08-31  53.75975   -1.73278
#> 5                    15.6702  2022-08-03 2022-08-31  53.75975   -1.73278
#> 6                    26.4942  2022-11-30 2023-01-04  53.75975   -1.73278
#> 7                    27.5438  2022-11-30 2023-01-04  53.75975   -1.73278
#> 8                    26.7238  2022-11-30 2023-01-04  53.75975   -1.73278
#> 9                    19.9096  2022-02-02 2022-03-02  53.75975   -1.73278
#> 10                   22.8042  2022-02-02 2022-03-02  53.75975   -1.73278
#> 11                   20.5984  2022-02-02 2022-03-02  53.75975   -1.73278
#> 12                   25.9038  2022-01-05 2022-02-02  53.75975   -1.73278
#> 13                   27.6996  2022-01-05 2022-02-02  53.75975   -1.73278
#> 14                   24.4770  2022-01-05 2022-02-02  53.75975   -1.73278
#> 15                   16.1376  2022-07-06 2022-08-03  53.75975   -1.73278
#> 16                   19.2208  2022-07-06 2022-08-03  53.75975   -1.73278
#> 17                   18.1548  2022-07-06 2022-08-03  53.75975   -1.73278
#> 18                   19.9834  2022-06-08 2022-07-06  53.75975   -1.73278
#> 19                   19.8276  2022-06-08 2022-07-06  53.75975   -1.73278
#> 20                   18.7206  2022-06-08 2022-07-06  53.75975   -1.73278
#> 21                   25.4856  2022-03-02 2022-03-30  53.75975   -1.73278
#> 22                   25.4364  2022-03-02 2022-03-30  53.75975   -1.73278
#> 23                   25.9940  2022-03-02 2022-03-30  53.75975   -1.73278
#> 24                   18.5976  2022-05-04 2022-06-08  53.75975   -1.73278
#> 25                   19.0650  2022-05-04 2022-06-08  53.75975   -1.73278
#> 26                   16.5312  2022-05-04 2022-06-08  53.75975   -1.73278
#> 27                   24.9280  2022-11-02 2022-11-30  53.75975   -1.73278
#> 28                   26.9288  2022-11-02 2022-11-30  53.75975   -1.73278
#> 29                   24.4770  2022-11-02 2022-11-30  53.75975   -1.73278
#> 30                   21.0330  2022-09-28 2022-11-02  53.75975   -1.73278
#> 31                   21.7382  2022-09-28 2022-11-02  53.75975   -1.73278
#> 32                   22.0744  2022-09-28 2022-11-02  53.75975   -1.73278
#> 33                   16.2114  2022-08-31 2022-09-28  53.75975   -1.73278
#> 34                   18.5156  2022-08-31 2022-09-28  53.75975   -1.73278
#> 35                   13.6858  2022-03-30 2022-05-04  53.75981   -1.76838
#> 36                   10.4222  2022-03-30 2022-05-04  53.75981   -1.76838
#> 37                   12.5460  2022-08-03 2022-08-31  53.75981   -1.76838
#> 38                   13.0298  2022-08-03 2022-08-31  53.75981   -1.76838
#> 39                   12.4148  2022-08-03 2022-08-31  53.75981   -1.76838
#> 40                   19.2044  2022-11-30 2023-01-04  53.75981   -1.76838
#> 41                   17.8678  2022-11-30 2023-01-04  53.75981   -1.76838
#> 42                   16.9084  2022-11-30 2023-01-04  53.75981   -1.76838
#> 43                   10.9880  2022-02-02 2022-03-02  53.75981   -1.76838
#> 44                   11.8162  2022-02-02 2022-03-02  53.75981   -1.76838
#> 45                   10.2172  2022-02-02 2022-03-02  53.75981   -1.76838
#> 46                   15.9408  2022-01-05 2022-02-02  53.75981   -1.76838
#> 47                   14.5304  2022-01-05 2022-02-02  53.75981   -1.76838
#> 48                   15.3996  2022-01-05 2022-02-02  53.75981   -1.76838
#> 49                    9.1184  2022-07-06 2022-08-03  53.75981   -1.76838
#> 50                    9.6350  2022-07-06 2022-08-03  53.75981   -1.76838
#> 51                    9.2824  2022-07-06 2022-08-03  53.75981   -1.76838
#> 52                    6.8798  2022-06-08 2022-07-06  53.75981   -1.76838
#> 53                    7.1094  2022-06-08 2022-07-06  53.75981   -1.76838
#> 54                    6.7978  2022-06-08 2022-07-06  53.75981   -1.76838
#> 55                   25.6086  2022-03-02 2022-03-30  53.75981   -1.76838
#> 56                   24.2966  2022-03-02 2022-03-30  53.75981   -1.76838
#> 57                   23.6242  2022-03-02 2022-03-30  53.75981   -1.76838
#> 58                    9.5366  2022-05-04 2022-06-08  53.75981   -1.76838
#> 59                    8.1918  2022-05-04 2022-06-08  53.75981   -1.76838
#> 60                    9.1348  2022-05-04 2022-06-08  53.75981   -1.76838
#> 61                   17.6464  2022-11-02 2022-11-30  53.75981   -1.76838
#> 62                   17.4578  2022-11-02 2022-11-30  53.75981   -1.76838
#> 63                   20.5410  2022-11-02 2022-11-30  53.75981   -1.76838
#> 64                   12.8002  2022-09-28 2022-11-02  53.75981   -1.76838
#> 65                   12.9314  2022-09-28 2022-11-02  53.75981   -1.76838
#> 66                   12.9560  2022-09-28 2022-11-02  53.75981   -1.76838
#>    .sample_id  .value
#> 1           4 15.9818
#> 2           4 17.1790
#> 3           4 13.4726
#> 4           8 18.7124
#> 5           8 15.6702
#> 6          12 26.4942
#> 7          12 27.5438
#> 8          12 26.7238
#> 9           2 19.9096
#> 10          2 22.8042
#> 11          2 20.5984
#> 12          1 25.9038
#> 13          1 27.6996
#> 14          1 24.4770
#> 15          7 16.1376
#> 16          7 19.2208
#> 17          7 18.1548
#> 18          6 19.9834
#> 19          6 19.8276
#> 20          6 18.7206
#> 21          3 25.4856
#> 22          3 25.4364
#> 23          3 25.9940
#> 24          5 18.5976
#> 25          5 19.0650
#> 26          5 16.5312
#> 27         11 24.9280
#> 28         11 26.9288
#> 29         11 24.4770
#> 30         10 21.0330
#> 31         10 21.7382
#> 32         10 22.0744
#> 33          9 16.2114
#> 34          9 18.5156
#> 35         51 13.6858
#> 36         51 10.4222
#> 37         55 12.5460
#> 38         55 13.0298
#> 39         55 12.4148
#> 40         59 19.2044
#> 41         59 17.8678
#> 42         59 16.9084
#> 43         49 10.9880
#> 44         49 11.8162
#> 45         49 10.2172
#> 46         48 15.9408
#> 47         48 14.5304
#> 48         48 15.3996
#> 49         54  9.1184
#> 50         54  9.6350
#> 51         54  9.2824
#> 52         53  6.8798
#> 53         53  7.1094
#> 54         53  6.7978
#> 55         50 25.6086
#> 56         50 24.2966
#> 57         50 23.6242
#> 58         52  9.5366
#> 59         52  8.1918
#> 60         52  9.1348
#> 61         58 17.6464
#> 62         58 17.4578
#> 63         58 20.5410
#> 64         57 12.8002
#> 65         57 12.9314
#> 66         57 12.9560
#>  [ reached 'max' / getOption("max.print") -- omitted 11207 rows ]