Diffusion Tube Accuracy

Testing diffusion tube accuracy. Coded methods are all currently based on the comparison of aligned diffusion tube (DT) and reference method (continuous analyser, CA) measurements.

testTubeAccuracy(
  data,
  data.ref = NULL,
  tube = ".value",
  ref = "no2",
  method = 1,
  max.distance = 10,
  nearest.only = FALSE,
  show = c("plot", "summary.report"),
  ...
)

Arguments

data

Data source, typically a data.frame or similar, containing data-series of conventionally formatted DT records.

data.ref

Reference data source, typically a data.frame or similar, containing data-series of conventionally formatted CA records not already aggregated, time aligned or added to data.

tube

The name of the data-series (in data) to test, typically DT NO2 concentrations (in ug/m3).

ref

The name of the data-series (in data.ref) to compare with tube.

method

The method to apply when aggregating ref (and data.ref) to compare with tube (and data). There is currently one method, set up to run with openair-format data.ref data sets. This method aggregates and averages data.ref at the sampling resolution of the tube data set, data, as identified from tags built using tagTube methods, before merging tube and reference data sets. See below or ?tagTube for further details.

max.distance

the maximum distance (m) between tube and reference method locations when pairing diffusion tube and reference measurements for comparison.

nearest.only

A logical (default FALSE); if TRUE only the nearest tube/ref pairings are retained.

show

The outputs to show when returning results, by default the plot and summary report.

...

Additional arguments, currently passed to tagTube to check tube tags, and tubePlot to handle generic plotting arguments.

Value

testTubeAccuracy returns a list of test results, typically data, the match (tube and reference) measurement pairs less than max.distance apart based on tagged sampling dates and locations, all, all date-based measurement pairs regardless of distance apart, plot, and results lookup table and report.

Details

Both sampling dates and locations are identified/tagged using tagTube methods. tagTubeStartEnd tags sampling dates and tagTubeLatLon tags locations.

NOTE: testTubeAccuracy attempts to tag data using default tagTube setting if it has not already been tagged, but it is probably better to tag and check all data yourself before running more detailed analysis.

testTubeAccuracy also uses tubePlot to handle common plotting arguments, e.g. X and Y axes label handling. Plot conditioning arguments like group and facet are also tracked by testTubeAccuracy itself and used to subset data when comparing tube and near-ref measurements.

When merging data and data.ref, testTubeAccuracy adds a '.ref' suffix to data columns from data.ref, so it you want to facet or group data by a column in data.ref, you need to include the suffix when naming it.

Examples

if (FALSE) { # \dontrun{
# importAURN require internet access
aurn <- openair::importAURN(site="bdma", year=2022:2025,
                            meta=TRUE)
testTubeAccuracy(dt.brd, data.ref=aurn,
                 group=".year", col=".year")
} # }