Diffusion Tube Precision

Testing diffusion tube precision. Coded methods are all currently based on the comparison of co-located replicate diffusion tube (DT) samples.

testTubePrecision(
  data,
  tube = "measurement",
  n = 3,
  method = 2,
  show = c("plot", "summary.report"),
  ...
)

Arguments

data

Data source, typically a data.frame or similar, containing data-series of conventionally formatted diffusion tube records.

tube

The name of the data-series (in data) to test, typically DT NO2 concentrations (in ug/m3).

n

The test compares replicate samples collected from the same location at the same time. n sets the number of replicates required, by default 3 for triplicate measurements.

method

The method to apply when testing tube data. See relevant details below for further information.

show

The outputs to show when returning results, by default the plot and summary report.

...

Additional arguments, currently passed to tagTube to check tube tags, and ggplotTubeShell to handle generic plotting arguments.

Value

testTubePrecision returns a list of test results, typically data all n-replicate measurement sets identified using tagTube, plot, results lookup table and; the (full) report.

Details

testTubePrecision uses tagTube to identify replicate samples, if not already tagged, then applies one of the following methods to estimate the precision of cases with n replicates:

method = 1 uses quantile(x, 0.05, 0.95)

method = 2 uses mean(x) +/- qt(0.05/2, n(x)-1), lower.tail=FALSE) * sd(x)/sqrt(n(x))

method = 3 used mean(x) +/- 1.96 * sd(x)/sqrt(n(x)))

method = 4 used mean(x) +/- qnorm(0.975) * sd(x)

NOTE: testTubePrecision attempts to tag data using default tagTube setting if it has not already been tagged, but it is probably better to tag and check all data yourself before running more detailed analysis.

testTubePrecision also uses ggplotTubeShell handles common plotting arguments, e.g. X and Y axes label handling. Plot conditioning arguments like group and facet are also tracked by testTubePrecision itself and used to subset data when testing precision.