Diffusion Tube Meta Data

Testing diffusion tube meta-data. Coded methods are all currently based on the visualisation of meta-data-like behaviour.

testTubeMeta(data, x = NULL, by = NULL, ...)

Arguments

data

Data source, typically a data.frame or similar, containing data-series of diffusion tube records.

x

(Optional) The name of one or more data-series in data or expressions to be evaluated, supplied as character strings. If not supplied, this defaults to all data-series in code

by

(Optional) The name of one or more data-series or tags to be used as a case identifier for meta information. If not supplied, this defaults to the all.data, sample, location and date properties of the data set tags. See Details below.

...

additional arguments, currently ignored.

Value

testTubeMeta generates a plot of meta-like behaviour. It is a standard ggplot2 output that can be modified using standard ggplot2 methods.

Details

testTubeMeta generates a plot of meta-like behaviour. It is currently in-development but is intended to be used in the form:

testTubeMeta(dt.data)

testTubeMeta attempts to identify data-series that look like meta information in the supplied data. By its nature, meta-data is expected to unique at a specific level of aggregation. For example, all samples collected at one site should have the same latitude, longitude, sample type, etc regardless of their date of collection. By default, testTubeMeta looks for data-series that are meta-like at four levels of aggregation: all.data indicating that all records are identical, sample indicating that all records for a given location and date combination are identical; location indicating that all records for a given location are identical; and date indicating that all records for a given date are identical. Any of these that score 100 percentage indicate meta-like behaviour at the associated level of aggregation. Although, a meta-like does not prove a data-series is meta-data, less than 100 percent scores for a data-series that is expected to be meta-data, would indicate some issues with that data-series.

Although the function default is to look for known sample identifiers, this grouping term is set using the addition argument by. The function then extracts all data-series with only one unique non-NA value. By default, the function test all non-grouping data-series, but similarly testing/extraction can be limited to specific cases using x. Please remeber that this function extracts data-series that at the by grouping-level looks like meta information.

See also

Examples

if (FALSE) { # \dontrun{
# (>5 secs during package build)
testTubeMeta(dt.brd)
} # }