Diffusion Tube Pro-processing

Pre-processing diffusion tube (DT) data for use with DTEval. Coded methods to standardise DT data collected using different DT sampling procedures or different logging conventions. Tagging makes a version of your tube information in a format DTEval can more easily work with, while retaining your original data in the format you are familiar with.

All tag... functions are intended to be run in the form:

tagged.data <- tagTube(data, ...)

(where ... refers to any additional tagging arguments)

tagTube(data, ...)

tagTubeDates(data, method = -1, force = FALSE, ...)

tagTubeLatLon(data, method = -1, force = FALSE, ...)

tagTubeSiteID(data, method = -1, force = FALSE, ...)

Arguments

data

Data source, typically a data.frame or similar, containing data-series of diffusion tube records.

...

Additional arguments can be used to assign data sources if these are not automatically identified by the tag... functions. See below for further details.

method

The method to apply when testing or modifying data. The default (method=-1) means try all methods in order until one works. See below for further details on different function methods.

force

(logical) By default, these functions do not overwrite existing tags if they are already present. The option force=TRUE forces them to rebuild the requested tag using current setting.

Value

These functions typically return the supplied data modified as requested.

Details

All tag... functions allow you to change information sources when adding tags. DTEval has defaults it is expects, but your files might be different. So, you can identify the specific data columns you would like used using additional arguments, e.g.:

year to assign the calendar year

month to assign the calendar month

start and end to assign the individual DT sampling period directly (start deployment and end collection, respectively). start and stop are assumed to be provided in conventional 'YYYY-MM-DD' format, but this can also be modified using data.format and strptime formatting instructions.

lat and lon to assign latitude and longitude, respectively.

So, for example:

tagTubeLatLon(data, lat="lat", lon="long")

tells tagTubeLatLon to use the data columns named 'lat' and 'long' as the sources when tagging latitude and longitude, respectively.

Function Methods:

By default, these methods try all available methods in order, until they find a method that works and they run out methods. If you want to use a specific method, or method order, just specific them, e.g. to tag tube dates using first method 3 then 2:

tagTubeDates(data, method=3:2)

tagTubeDates attempts to build Date class time-stamps for sampling dates. It currently has three methods: method=1 use month and year records to build a nominal calendar monthly sampling record, i.e., first to last day of each month; method=2 use month and year records to build a sampling calendar based on the DEFRA diffusion tube sampling calendar (see dt.calendar); and, method=3 to build directly using user-assigned sampling start and end dates. If sampling dates are built, these are added to data as columns .start_date and .end_date.

tagTubeLatLon identifies sample location. It currently uses one method: method=1 assigning all locations based on identified latitude and longitude. These are then tagged as .latitude and .longitude, respectively.

tagTubeSiteID identifies replicate samples. It currently uses one method: method=1 assigning all same-location-and-time DTs as replicates, and is added to data as .site_id.

tagTube is a wrapper that runs tagTubeDates, tagTubeLatLon and tagTubeiteID. Other DTEval functions typically pass your data to tagTube to check for tags and add them if missing. If you are using tagTube or one of these on untagged data and need to specific individual methods for each tagTube... function, prefix the method request, e.g. to use method 2 for tabTubeDates but leave the other functions using default setting when applying tagTube, use:

tagTube(data, dates.method=2)